changed secure text entry from dots to stars, fixed bug when a float number did not show the fractional part

+++ app/src/main/java/com/dandelion/controltext/ui/components/NumericField.kt
@@ -51,7 +51,7 @@ fun NumericField(label: String, initValue: Int, onValueChange: (Int) -> Unit) {
 @Composable
 fun FloatNumericField(label: String, initValue: Float, onValueChange: (Float) -> Unit) {
     var value by remember(key1 = onValueChange) {
-        mutableStateOf((initValue.toInt()).toString())
+        mutableStateOf(initValue.toString())
     }
 
     Row(
